requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.6
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1689373676338/work
cycler==0.11.0
datasets==2.14.3
dill==0.3.7
docker-pycreds==0.4.0
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.32
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
huggingface-hub==0.16.4
idna @ file:///croot/idna_1666125576474/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
kiwisolver==1.4.4
loguru==0.7.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.2
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///croot/numpy_and_numpy_base_1687466208449/work
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pathtools==0.1.2
Pillow==9.4.0
platformdirs==3.10.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.0.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests @ file:///croot/requests_1690400202158/work
safetensors==0.3.1
seaborn==0.12.2
sentry-sdk==1.29.2
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sympy @ file:///croot/sympy_1668202399572/work
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1686163155763/work
wandb==0.15.8
wcwidth==0.2.6
xxhash==3.3.0
yarl==1.9.2
